# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API built with Express."
  },
  "paths": {
  "/": {
    "get": {
      "summary": "API information",
      "responses": {
        "200": {
          "description": "Returns the API name and version."
        }
      }
    }
  },

  "/health": {
    "get": {
      "summary": "Health check",
      "responses": {
        "200": {
          "description": "Returns the server status."
        }
      }
    }
  },

  "/tasks": {
    "get": {
      "summary": "Get all tasks",
      "responses": {
        "200": {
          "description": "Returns all tasks."
        }
      }
    },

    "post": {
      "summary": "Create a new task",
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "example": {
              "title": "Buy milk"
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "Task created successfully."
        },
        "400": {
          "description": "Invalid input."
        }
      }
    }
  },

  "/tasks/{id}": {
    "get": {
      "summary": "Get a task by ID",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "Task found."
        },
        "404": {
          "description": "Task not found."
        }
      }
    },

    "put": {
      "summary": "Update a task",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "example": {
              "title": "Buy groceries",
              "done": true
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Task updated."
        },
        "400": {
          "description": "Invalid request."
        },
        "404": {
          "description": "Task not found."
        }
      }
    },

    "delete": {
      "summary": "Delete a task",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "responses": {
        "204": {
          "description": "Task deleted."
        },
        "404": {
          "description": "Task not found."
        }
      }
    }
  }
}
}